requirements.txt
Django>=4.2
djangorestframework
djoser
drf-social-oauth2
drf-yasg
reportlab
openpyxl
Pillow
